requirements.txt
asgiref==3.4.1
certifi==2021.10.8
Django==3.2.8
django-filter==21.1
djangorestframework==3.12.4
djangorestframework-simplejwt==5.0.0
psycopg2==2.9.1
PyJWT==2.3.0
pytz==2021.3
sentry-sdk==1.4.3
sqlparse==0.4.2
urllib3==1.26.7
